+++ changelog.md
@@ -1,4 +1,11 @@
-# Pelion E2E Edge Python Test Suite Change log
+# Izuma E2E Edge Python Test Suite Change log
+
+## 1.1.0  2022-Sep-01
+- Izuma branding changes.
+- Remove EU and JP config templates.
+- Fix license link.
+- Fix Jenkinsfile link.
+- Update requirements.txt to work better with Python 3.10.
 
 ## 1.0.0  2021-Mar-04
 - Initial release of Pelion E2E Edge Python Test Suite